+++ src/Model/ClientList.php
@@ -5,9 +5,8 @@ declare(strict_types=1);
 namespace amcintosh\FreshBooks\Model;
 
 use Spatie\DataTransferObject\Attributes\CastWith;
-use Spatie\DataTransferObject\Attributes\MapFrom;
 use Spatie\DataTransferObject\Casters\ArrayCaster;
-use Spatie\DataTransferObject\DataTransferObject;
+use amcintosh\FreshBooks\Model\AccountingList;
 use amcintosh\FreshBooks\Model\Client;
 
 /**
@@ -16,18 +15,9 @@ use amcintosh\FreshBooks\Model\Client;
  * @package amcintosh\FreshBooks\Model
  * @link https://www.freshbooks.com/api/clients
  */
-class ClientList extends DataTransferObject
+class ClientList extends AccountingList
 {
     public const RESPONSE_FIELD = 'clients';
-
-    public int $page;
-
-    public int $pages;
-
-    #[MapFrom('per_page')]
-    public int $perPage;
-
-    public int $total;
 
     #[CastWith(ArrayCaster::class, itemType: Client::class)]
     public array $clients;